Resolved Issues: #340:  Reports - Pass and Fail percentage displayed in the browser report is incorrect.
#299 - comments : If there is only one test case in a scenario, on click 
of the scenario name, the alignment of details is not proper.

#110 - ICE_Screen : When "Select All" checkbox is selcted in "Filters" 
panel, screen goes into unresponsive mode for approx 60 seconds 
(integrated progress loader)

#96 - ICE_Test Case - Not able to save the test case if the 
objectname/input/output column contains ' (single quote).

Font format for specific repors is changed to "Latoweb" which is
used in ICE screens.

Two digit format for Month and Date in the startDate and endDate of
reports screen is fixed.
modif